api/uploads

+++ src/app/api/uploads/route.js
@@ -1,10 +1,10 @@
 import { NextResponse } from "next/server";
 import { getUser } from "@/lib/api";
 import {
-  getUploadsWhere,
+  getUploadsWithAllWhere,
+  getUploadById,
   getUploadByFileHash,
   createUpload,
-  getUploadById,
   deleteUploadWhere,
 } from "@/lib/uploads";
 import { validatePostRequest, constructPostData } from ".";
@@ -28,7 +28,7 @@ export const GET = async (request) => {
       where.uploads.id = [];
     }
 
-    const uploads = await getUploadsWhere(where);
+    const uploads = await getUploadsWithAllWhere(where);
     return NextResponse.json(
       { success: true, data: uploads, message: null },
       { status: 200 }
@@ -55,7 +55,7 @@ export const POST = async (request) => {
 
     if (!canCreateUpload) {
       return NextResponse.json(
-        { success: false, data: null, message: "Autorisations insuffisantes" },
+        { success: false, data: null, message: "Permissions insuffisantes" },
         { status: 403 }
       );
     }
@@ -79,7 +79,8 @@ export const POST = async (request) => {
       },
       { status: 201 }
     );
-  } catch {
+  } catch (e) {
+    console.log(e);
     return NextResponse.json(
       { success: false, data: null, message: "Erreur interne du serveur" },
       { status: 500 }
@@ -90,45 +91,82 @@ export const POST = async (request) => {
 export const DELETE = async (request) => {
   try {
     const { searchParams } = new URL(request.url);
-    const id = searchParams.get("id");
+    const ids = searchParams.getAll("id");
 
     const { id: userId, permissions = [] } = await getUser();
     const canDeleteAllUploads = permissions.includes("CAN_DELETE_ALL_UPLOADS");
     const canDeleteOwnUploads = permissions.includes("CAN_DELETE_OWN_UPLOADS");
 
-    const where = { uploads: {} };
-    if (canDeleteAllUploads) {
-      where.uploads.id = id;
-    } else if (canDeleteOwnUploads) {
-      where.uploads.id = id;
-      where.uploads.user_id = userId;
-    } else {
-      where.uploads.id = null;
-    }
+    if (ids.length === 1) {
+      const upload = await getUploadById(ids[0]);
+      if (!upload) {
+        return NextResponse.json(
+          { success: false, data: [], message: "Téléversement introuvable" },
+          { status: 404 }
+        );
+      }
 
-    const uploadId = await deleteUploadWhere(where);
-    if (!uploadId) {
+      const isNotAllowed =
+        !canDeleteAllUploads &&
+        (!canDeleteOwnUploads || upload.user_id !== userId);
+
+      if (isNotAllowed) {
+        return NextResponse.json(
+          { success: false, data: [], message: "Permissions insuffisantes" },
+          { status: 403 }
+        );
+      }
+
+      const where = { uploads: { id: ids[0] } };
+      const deletedUploadId = await deleteUploadWhere(where);
+
       return NextResponse.json(
         {
-          success: false,
-          data: null,
-          message: "Autorisation insuffisantes ou téléversement introuvable",
+          success: true,
+          data: [deletedUploadId],
+          message: "Téléversement supprimé avec succès",
         },
-        { status: 500 }
+        { status: 200 }
       );
     }
 
+    const deletedUploadIds = [];
+    for (const id of ids) {
+      const where = { uploads: {} };
+      if (canDeleteAllUploads) {
+        where.uploads.id = id;
+      } else if (canDeleteOwnUploads) {
+        where.uploads.id = id;
+        where.uploads.user_id = userId;
+      } else {
+        where.uploads.id = null;
+      }
+      await deleteUploadWhere(where)
+        .then((id) => {
+          if (id) deletedUploadIds.push(id);
+        })
+        .catch(() => null);
+    }
+
+    let message, status;
+    if (deletedUploadIds.length === ids.length) {
+      message = "Tous les téléversements ont été supprimés avec succès";
+      status = 200;
+    } else if (deletedUploadIds.length === 0) {
+      message = "Aucun téléversement n'a été supprimé";
+      status = 404;
+    } else {
+      message = "Certains téléversements ont été supprimés avec succès";
+      status = 200;
+    }
+
     return NextResponse.json(
-      {
-        success: true,
-        data: uploadId,
-        message: "Téléversement a été supprimé avec succès",
-      },
-      { status: 200 }
+      { success: true, data: deletedUploadIds, message },
+      { status }
     );
-  } catch (e) {
+  } catch {
     return NextResponse.json(
-      { success: false, data: null, message: "Erreur interne du serveur" },
+      { success: false, data: [], message: "Erreur interne du serveur" },
       { status: 500 }
     );
   }
